requirement.txt
boto3
finnhub-python